User settings has been updated and added private user management

+++ app.js
@@ -106,7 +106,8 @@ mongoose.connect(config.MONGODB, {
 
           if (user) {
             // Allowing admin and registered user to remove his account
-            if (message.member.roles.some(r => config.ADMIN_ROLES.includes(r.name))
+            if ((message.guild
+              && message.member.roles.some(r => config.ADMIN_ROLES.includes(r.name)))
               || user.discordId === message.author.id) {
               User.deleteOne({ name: username })
                 .then(() => {
@@ -132,7 +133,8 @@ mongoose.connect(config.MONGODB, {
           message.reply('**ERROR:** URL is not valid.');
         } else {
           // Checking if the message author has the privilege
-          if (!message.member.roles.some(r => config.ADMIN_ROLES.includes(r.name))) {
+          if (!message.guild
+            || !message.member.roles.some(r => config.ADMIN_ROLES.includes(r.name))) {
             return;
           }
 
@@ -193,7 +195,8 @@ mongoose.connect(config.MONGODB, {
           message.reply('URL is not valid.');
         } else {
           // Checking if the message author has the privilege
-          if (!message.member.roles.some(r => config.ADMIN_ROLES.includes(r.name))) {
+          if (!message.guild
+            || !message.member.roles.some(r => config.ADMIN_ROLES.includes(r.name))) {
             return;
           }
 
@@ -240,28 +243,40 @@ mongoose.connect(config.MONGODB, {
 
       // Pause Command
       case 'set': {
-        let [username, prop, boolean] = splitMessage.slice(1);
+        let [username, prop, value] = splitMessage.slice(1);
 
         if (!username) {
           message.reply('**ERROR:** Username is required.');
-        } else if (!['pause', 'comment'].includes(prop)) {
+        } else if (!['pause', 'comment', 'heal'].includes(prop)) {
           message.reply('**ERROR:** Unrecognized settings property.');
+        } else if (prop === 'heal' && !['on', 'off', 'only'].includes(value)) {
+          message.reply('**ERROR**: Property `heal` can only have one of `on, off, or only` as value.');
+        } else if (!value) {
+          message.reply(`**ERROR**: A settings value is required for \`${prop}\`.`);
         } else {
           username = username.toLowerCase();
           prop = prop.toLowerCase();
+
+          if (prop !== 'heal') {
           // Converting string to boolean, if sting is true output is true else it is false
-          boolean = boolean.toLowerCase() === 'true';
+            value = value.toLowerCase() === 'true';
+          } else {
+            value = value.toLowerCase();
+          }
 
           const user = await User.findOne({ name: username });
 
           if (user) {
             // Allowing admin and the registered user to update the account
-            if (message.member.roles.some(r => config.ADMIN_ROLES.includes(r.name))
+            if ((message.guild
+              && message.member.roles.some(r => config.ADMIN_ROLES.includes(r.name)))
               || user.discordId === message.author.id) {
               if (prop === 'pause') {
-                user.paused = boolean;
+                user.settings.paused = value;
               } else if (prop === 'comment') {
-                user.comment = boolean;
+                user.settings.comment = value;
+              } else if (prop === 'heal') {
+                user.settings.heal = value;
               }
 
               await user.save()
@@ -294,6 +309,13 @@ mongoose.connect(config.MONGODB, {
         break;
       }
 
+      case 'private': {
+        message.delete();
+        message.member.send('Hey, how can I help you?');
+
+        break;
+      }
+
       case 'help': {
         const help = stripIndents`
         **Available Commands:**
@@ -313,12 +335,15 @@ mongoose.connect(config.MONGODB, {
         **\`${config.PREFIX}delete username\`**
           *username - Registered Steem username.*
 
-        **\`${config.PREFIX}set prop boolean\`**
-          *prop - It can be pause or comment.*
-          *boolean - Set true to pause/comment or false to resume/no comment.*
+        **\`${config.PREFIX}set prop value\`**
+          *prop - It can be pause/comment/heal.*
+          *value - Set to true to pause/comment or false to resume/no comment. Set to on, off, or only for heal.*
 
         **\`${config.PREFIX}stats\`**
           *Displays overall statistics of the trail.*
+
+        **\`${config.PREFIX}private\`**
+          *To privately add/remove accounts, and manage associated settings.*
         `;
 
         message.channel.send(help);
@@ -334,7 +359,7 @@ mongoose.connect(config.MONGODB, {
 })();
 
 process.on('uncaughtException', (err) => {
-  console.log(err);
+  console.log(err.message);
 });
 
 process.on('SIGINT', () => {
